requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==4.4.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
multidict==6.0.4
numpy==1.25.2
openai==0.27.8
packaging==23.1
Pillow==10.0.0
proglog==0.1.10
pydub==0.25.1
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.31.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2